[metasrv] refactor: move test supporting utils into metasrv/raft/state_machine/testing/
